Add Email field to User model and update database schema
Added a required `Email` property to the `User` model with a
default value of `default!`. Updated `ApplicationDbContextModelSnapshot.cs`
to reflect the new `Email` property in the `Users` table.

Created migration `20251008070648_ModifyUser` to add the `Email`
column to the database schema. Included `Up` and `Down` methods
to handle schema changes. Generated the corresponding migration
designer file to update the `BuildTargetModel` method with the
new `Email` property.